Gui improvements: Cables, File Menu, Notifications Bar, Add Module

+++ firmware/src/gui/pages/page_args.hh
@@ -12,7 +12,7 @@ struct PageArguments {
 	std::optional<PatchLocHash> patch_loc_hash{};
 	std::optional<uint16_t> module_id{};
 	std::optional<uint16_t> mappedknob_id{};
-	std::optional<uint16_t> view_knobset_id{};
+	std::optional<uint32_t> view_knobset_id{};
 	std::optional<ElementCount::Counts> element_counts{};
 	std::optional<ElementCount::Indices> element_indices{};
